move: ProfileGuest, ProfileLoggedIn, SideMenuItem 파일 위치 변경 (#126)
